+++ modules/menus/media/index.ts
@@ -1,6 +1,6 @@
 import Window from 'types/widgets/window.js';
 import DropdownMenu from '../shared/dropdown/index.js';
-import { Media } from './media.js';
+import { Media } from './components/index.js';
 import { Attribute, Child } from 'lib/types/widget.js';
 import options from 'options.js';
